08-connect-with-mysql/pyproject.toml
[project]
name = "08-connect-with-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "mysql-connector-python>=9.3.0",
    "python-dotenv>=1.1.1",
]
